agent/requirements.txt
# LiveKit Agents SDK
livekit-agents[deepgram,silero,openai]

# Maya1 TTS - FREE open source with emotional voices
transformers>=4.40.0
torch>=2.0.0
snac
soundfile
numpy

# Supabase for CRM data
supabase>=2.0.0

# Other dependencies
python-dotenv>=1.0.0
httpx>=0.27.0
